layout/base/tests/test_bug582181-1.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=582181
-->
<head>
  <title>Test for Bug 582181</title>
  <meta charset="utf-8">
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="test()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=582181">Mozilla Bug 582181</a>
<p id="display"></p>
<div id="content" dir="rtl">
<textarea rows="4" id="testInput">فارسی
[[en:Farsi]]</textarea>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 582181 **/

SimpleTest.waitForExplicitFinish();

function test() {
  var textInput = $("testInput");
  var s1, s2, s3, equal, str1, str2;

  s1 = snapshotWindow(window);

  textInput.focus();
  synthesizeKey("VK_DOWN", { });
  synthesizeKey("VK_DOWN", { });
  synthesizeKey("VK_RETURN", { });
  textInput.blur();
  s2 = snapshotWindow(window);

  [equal, str1, str2] = compareSnapshots(s1, s2, true);
  ok(equal, "enter after text shouldn't change rendering: expected " +
     str1 + " but got " + str2);

  textInput.focus();
  synthesizeKey("VK_BACK_SPACE", { });
  textInput.blur();
  s3 = snapshotWindow(window);

  [equal, str1, str2] = compareSnapshots(s1, s3, true);
  ok(equal, "backspace shouldn't change rendering: expected " + str1 + 
     " but got " + str2);

  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
